# os/clearenv: optimise cgo clearenv