For programs with very large environments, calling unsetenv(3) for each environment variable can be fairly expensive because of CGo overhead, but clearenv(3) is much faster. The only thing we have to track is whether GODEBUG is being unset by the operation, which can be done very quickly without resorting to doing unsetenv(3) for every variable.

This change makes os.Clearenv() ~30% faster when run an environment with 1000 environment variable set (and gets better with more environment variables):

    cpu: AMD Ryzen 7 7840U w/ Radeon  780M Graphics
           │   old.txt   │               new.txt               │
           │   sec/op    │   sec/op     vs base                │
    Env-16   213.6µ ± 1%   185.1µ ± 3%  -13.32% (p=0.000 n=10)

(The time taken to set up 1000 variables in this benchmark is 117.2µs, so while the naive timing is 13% the speedup for os.Clearenv() itself the speedup is from 96.4µs to 67.9µs which is ~30%.)